refactor(test): extract shared extractBlock helper, throw on missing anchor
Review follow-up for #772: font-rendering.test.ts duplicated the
extractBlock CSS-block parser verbatim from neumorphic-themes.test.ts.
Move it to a shared src/test-utils.ts and import it from both files.

While extracting, make the helper throw with a descriptive message when
the anchor (or its opening brace) is missing, so anchor drift in
index.css fails loudly instead of silently matching an empty string.
